# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ke Klinik REST ful API Documentation",
        "description": "Description endpoint rest full api project ke klinik",
        "version": "1.0.0"
    },
    "paths": {
        "/api/appointment": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Add new Appointment",
                "description": "Endpoint to add new appointment with patient_id & diagnose_id",
                "operationId": "addAppointment",
                "requestBody": {
                    "description": "Request addAppointment needed patient_id & diagnose_id",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "diagnose_id"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "description": "validation field patient_id: required, integer",
                                        "type": "integer"
                                    },
                                    "diagnose_id": {
                                        "description": "validation field diagnose_id: required, integer",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Appointment add successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment add successfully"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": "201"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "3325"
                                                },
                                                "patient_id": {
                                                    "type": "integer",
                                                    "example": "3456"
                                                },
                                                "diagnose_id": {
                                                    "type": "integer",
                                                    "example": "3475"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-09 07:59:39"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-12-09 07:59:39"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "message field error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Route not found"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": "404"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/appointment/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get detail Appointment by id",
                "description": "Endpoint to get detail appointment with id appointment",
                "operationId": "getDetailAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id appointment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment add successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get detail appointmen successfully"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment not found"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": "404"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Edit Appointment",
                "description": "Endpoint to edit appointment with patient_id, diagnose_id & status",
                "operationId": "editAppointment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id appointment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request addAppointment needed patient_id, diagnose_id & status",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "patient_id",
                                    "diagnose_id",
                                    "status"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "description": "validation field patient_id: required, integer",
                                        "type": "integer"
                                    },
                                    "diagnose_id": {
                                        "description": "validation field diagnose_id: required, integer",
                                        "type": "integer"
                                    },
                                    "status": {
                                        "description": "validation field service: required, only 0 or 1",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment edit successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment edit successfully"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "message field error"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "{}"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Appointment not found"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": "404"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "[]"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
...
}